@@ -1,7 +1,7 @@
 {
   "name": "sqlstring",
   "description": "Simple SQL escape and format for MySQL",
-  "version": "2.3.1",
+  "version": "2.3.2",
   "contributors": [
     "Adri Van Houdt <adri.van.houdt@gmail.com>",
     "Douglas Christopher Wilson <doug@somethingdoug.com>",
@@ -21,9 +21,9 @@
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "eslint": "4.18.1",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "nyc": "10.3.2",
+    "eslint": "5.16.0",
+    "eslint-plugin-markdown": "1.0.2",
+    "nyc": "15.0.0",
     "urun": "0.0.8",
     "utest": "0.0.8"
   },
